Register vocab file by correct file name
It wasn't registering because the name didnt match the file in the vocab folder.

This gets it to register and recognize the command, but self.speak_dialog is looking for a template called 'the next streetcar...', so it throws an error. But it did make a request and get the time. I manually edited the file in vim on the device, so you may need to "git checkout ." before git pull.

I think you have to create a template file for speech.